requirements.txt
CouchDB
Django==1.4.1
Mako
Pyro
amqplib
anyjson
argparse
celery
configglue
configobj
distribute
django-celery
django-debug-toolbar
django-evolution
django-keyedcache
django-model-utils
django-picklefield
django-piston
django_extensions
django-registration
foolscap
fstab
gunicorn
httplib2
kombu
lazr.restfulclient
lazr.uri
ordereddict
pexpect
pika
pyOpenSSL
pycrypto
pyflakes
pyserial
python-dateutil
pytz
pyxdg
rdflib
simplejson
urlgrabber
wadllib
wsgiref
zope.interface
librabbitmq
six
sse
django-sse
